+++ layers/CMSSchema/packages/commentmeta/src/FieldResolvers/ObjectType/CommentObjectTypeFieldResolver.php
@@ -75,7 +75,7 @@ class CommentObjectTypeFieldResolver extends AbstractWithMetaObjectTypeFieldReso
                 $commentMetaTypeAPI = $this->getCommentMetaTypeAPI();
                 $allCommentMetaKeys = $commentMetaTypeAPI->getCommentMetaKeys($comment);
                 foreach ($allCommentMetaKeys as $key) {
-                    if (!$commentMetaTypeAPI->validateIsMetaKeyAllowed($key) || $commentMetaTypeAPI->isMetaKeyProtected($key)) {
+                    if (!$commentMetaTypeAPI->validateIsMetaKeyAllowed($key) || $commentMetaTypeAPI->isMetaKeyProtectedFromReading($key)) {
                         continue;
                     }
                     $metaKeys[] = $key;