+++ layers/CMSSchema/packages/meta/src/TypeAPIs/MetaTypeAPIInterface.php
@@ -8,6 +8,7 @@ interface MetaTypeAPIInterface
 {
     public function validateIsMetaKeyAllowed(string $key): bool;
     public function isMetaKeyProtected(string $key): bool;
+    public function isMetaKeyProtectedFromReading(string $key): bool;
     /**
      * @return string[]
      */